refactor(dccpeep): extract terminal cleanup passes
Move conservative JP-to-JR relaxation, constant sign-extension folding, and dead 16-bit reload cleanup into peep_pass_final.c. The main module retains scheduler ownership so terminal ordering remains explicit.

Validation: direct fixtures passed and all 316 raw assembly inputs produced byte-identical output; no performance baselines changed.
